pyproject.toml
[project]
name = "novelist"
version = "0.1.0"
description = "多Agent协同的小说创作系统"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
maintainers = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["ai", "novel", "writing", "llm", "autogen"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pyyaml>=6.0",
    "pyautogen==0.4.8.1",
    "python-dotenv>=0.19.0",
    "typing-extensions>=4.5.0",
    "aiohttp>=3.8.0",
    "jsonschema>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/novelist"
Documentation = "https://github.com/yourusername/novelist#readme"
Repository = "https://github.com/yourusername/novelist.git"
Issues = "https://github.com/yourusername/novelist/issues"

[project.scripts]
novelist = "novelist.__main__:run"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v -ra --showlocals"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
